"Man Feedback 2021-01-18 (#4)"

+++ ShoppingCart/ShoppingCart/Interfaces/IShoppingCartCalculator.cs
@@ -4,6 +4,6 @@ namespace ShoppingCart.Interfaces
 {
     public interface IShoppingCartCalculator
     {
-        double Total(ShoppingCartRequest request);
+        decimal Total(ShoppingCartRequest request);
     }
 }
